Use i32 for camera width and height

+++ Fussion/src/Fussion/Rendering/2D/Camera.cpp
@@ -4,9 +4,9 @@
 
 namespace Fussion
 {
-    Camera2D::Camera2D(f32 width, f32 height, f32 minClip, f32 maxClip)
-        : m_size(1.0f), m_aspect(width / height), m_min_clip(minClip), m_max_clip(maxClip), m_screenWidth(width),
-          m_screenHeight(height)
+    Camera2D::Camera2D(i32 width, i32 height, f32 minClip, f32 maxClip)
+        : m_size(1.0f), m_aspect(static_cast<f32>(width) / static_cast<f32>(height)), m_min_clip(minClip),
+          m_max_clip(maxClip), m_screenWidth(width), m_screenHeight(height)
     {
         update_projection_matrix();
     }
@@ -16,11 +16,11 @@ namespace Fussion
         m_projectionMatrix = glm::ortho(-m_aspect * m_size, m_aspect * m_size, -m_size, m_size, m_min_clip, m_max_clip);
     }
 
-    void Camera2D::resize(f32 width, f32 height)
+    void Camera2D::resize(i32 width, i32 height)
     {
         m_screenWidth = width;
         m_screenHeight = height;
-        m_aspect = width / height;
+        m_aspect = static_cast<f32>(width) / static_cast<f32>(height);
         update_projection_matrix();
     }
