+++ Fussion/src/Fussion/Rendering/2D/Camera.h
@@ -11,14 +11,14 @@ namespace Fussion
         f32 m_size{1.0f};
         f32 m_aspect{};
         f32 m_min_clip{}, m_max_clip{};
-        f32 m_screenWidth, m_screenHeight;
+        i32 m_screenWidth, m_screenHeight;
 
         void update_projection_matrix();
 
     public:
-        Camera2D(f32 width, f32 height, f32 minClip = -1.0f, f32 maxClip = 1.0f);
+        Camera2D(i32 width, i32 height, f32 minClip = -1.0f, f32 maxClip = 1.0f);
 
-        void resize(f32 width, f32 height);
+        void resize(i32 width, i32 height);
 
         void set_size(f32 size)
         {
